Moved general use functions to geomesa-utils/GeometryUtils.scala and modified TubeBuilder to use the changes
